Move all partitionId related functions to SegmentUtils
